+++ src/routes/(app)/+layout.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
   import Toast from "../Toast.svelte";
   import Alert from "$lib/components/Alert.svelte";
+  import Header from "./Header.svelte";
   import Footer from "./Footer.svelte";
-  import Navbar from "./Navbar.svelte";
   import { getLocale } from "$paraglide/runtime";
 
   const { data, children } = $props();
 </script>
 
 <div class="flex min-h-screen flex-col">
-  <Navbar notificationsPromise={data.notificationsPromise} isApp={data.isApp} />
+  <Header notificationsPromise={data.notificationsPromise} isApp={data.isApp} />
 
   <main class="flex min-h-0 flex-1 flex-col">
     {#each data.alerts as alert (alert.id)}
